+++ src/config/API_URL.js
@@ -1,0 +1,9 @@
+const IP_URL = 'http://localhost:8000/';
+const SERVICE_PATH = {
+    LOGIN: IP_URL + 'auth/employee/login',
+    REGISTER: IP_URL + 'auth/employee/register',
+    LOGOUT: IP_URL + 'auth/employee/logout'
+
+};
+
+export default SERVICE_PATH;
